+++ src/main/java/net/silentchaos512/gear/item/gear/GearAxeItem.java
@@ -134,11 +134,6 @@ public class GearAxeItem extends AxeItem implements GearDiggerTool {
     }
 
     @Override
-    public Component getName(ItemStack stack) {
-        return GearHelper.getDisplayName(stack);
-    }
-
-    @Override
     public void appendHoverText(ItemStack stack, TooltipContext tooltipContext, List<Component> tooltip, TooltipFlag flagIn) {
         GearClientHelper.addInformation(stack, tooltipContext, tooltip, flagIn);
     }