+++ src/main/java/net/silentchaos512/gear/item/gear/GearCrossbowItem.java
@@ -16,7 +16,6 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.item.component.ItemAttributeModifiers;
 import net.minecraft.world.level.Level;
-import net.silentchaos512.gear.SilentGear;
 import net.silentchaos512.gear.api.item.GearType;
 import net.silentchaos512.gear.api.item.GearRangedWeapon;
 import net.silentchaos512.gear.client.util.GearClientHelper;
@@ -104,11 +103,6 @@ public class GearCrossbowItem extends CrossbowItem implements GearRangedWeapon {
     @Override
     public int getEnchantmentValue(ItemStack stack) {
         return GearHelper.getEnchantmentValue(stack);
-    }
-
-    @Override
-    public Component getName(ItemStack stack) {
-        return GearHelper.getDisplayName(stack);
     }
 
     @Override