+++ src/main/java/net/silentchaos512/gear/item/gear/GearCurioItem.java
@@ -90,11 +90,6 @@ public class GearCurioItem extends Item implements GearItem {
     }
 
     @Override
-    public Component getName(ItemStack stack) {
-        return GearHelper.getDisplayName(stack);
-    }
-
-    @Override
     public void setDamage(ItemStack stack, int damage) {
         GearHelper.setDamage(stack, damage, super::setDamage);
     }