+++ src/main/java/net/silentchaos512/gear/item/gear/GearFishingRodItem.java
@@ -151,11 +151,6 @@ public class GearFishingRodItem extends FishingRodItem implements GearTool {
     }
 
     @Override
-    public Component getName(ItemStack stack) {
-        return GearHelper.getDisplayName(stack);
-    }
-
-    @Override
     public void setDamage(ItemStack stack, int damage) {
         GearHelper.setDamage(stack, damage, super::setDamage);
     }