+++ src/main/java/net/silentchaos512/gear/item/gear/GearHoeItem.java
@@ -93,11 +93,6 @@ public class GearHoeItem extends HoeItem implements GearDiggerTool {
     }
 
     @Override
-    public Component getName(ItemStack stack) {
-        return GearHelper.getDisplayName(stack);
-    }
-
-    @Override
     public void setDamage(ItemStack stack, int damage) {
         GearHelper.setDamage(stack, damage, super::setDamage);
     }