+++ src/main/java/net/silentchaos512/gear/item/gear/GearMaceItem.java
@@ -65,11 +65,6 @@ public class GearMaceItem extends MaceItem implements GearTool {
     }
 
     @Override
-    public Component getName(ItemStack stack) {
-        return GearHelper.getDisplayName(stack);
-    }
-
-    @Override
     public void setDamage(ItemStack stack, int damage) {
         GearHelper.setDamage(stack, damage, super::setDamage);
     }