+++ src/main/java/net/silentchaos512/gear/item/gear/GearPickaxeItem.java
@@ -161,11 +161,6 @@ public class GearPickaxeItem extends PickaxeItem implements GearDiggerTool {
     }
 
     @Override
-    public Component getName(ItemStack stack) {
-        return GearHelper.getDisplayName(stack);
-    }
-
-    @Override
     public void appendHoverText(ItemStack stack, TooltipContext tooltipContext, List<Component> tooltip, TooltipFlag flagIn) {
         GearClientHelper.addInformation(stack, tooltipContext, tooltip, flagIn);
     }