+++ src/main/java/net/silentchaos512/gear/item/gear/GearShearsItem.java
@@ -103,11 +103,6 @@ public class GearShearsItem extends ShearsItem implements GearTool {
     }
 
     @Override
-    public Component getName(ItemStack stack) {
-        return GearHelper.getDisplayName(stack);
-    }
-
-    @Override
     public void setDamage(ItemStack stack, int damage) {
         super.setDamage(stack, GearHelper.calcDamageClamped(stack, damage));
         if (GearHelper.isBroken(stack)) {