+++ src/main/java/net/silentchaos512/gear/item/gear/GearShieldItem.java
@@ -93,11 +93,6 @@ public class GearShieldItem extends ShieldItem implements GearItem {
     }
 
     @Override
-    public Component getName(ItemStack stack) {
-        return GearHelper.getDisplayName(stack);
-    }
-
-    @Override
     public void setDamage(ItemStack stack, int damage) {
         super.setDamage(stack, GearHelper.calcDamageClamped(stack, damage));
         if (GearHelper.isBroken(stack)) {