+++ src/main/java/net/silentchaos512/gear/item/gear/GearShovelItem.java
@@ -123,11 +123,6 @@ public class GearShovelItem extends ShovelItem implements GearDiggerTool {
     }
 
     @Override
-    public Component getName(ItemStack stack) {
-        return GearHelper.getDisplayName(stack);
-    }
-
-    @Override
     public void appendHoverText(ItemStack stack, TooltipContext tooltipContext, List<Component> tooltip, TooltipFlag flagIn) {
         GearClientHelper.addInformation(stack, tooltipContext, tooltip, flagIn);
     }