+++ src/main/java/net/silentchaos512/gear/item/gear/GearTridentItem.java
@@ -81,11 +81,6 @@ public class GearTridentItem extends TridentItem implements GearWeapon {
     }
 
     @Override
-    public Component getName(ItemStack stack) {
-        return GearHelper.getDisplayName(stack);
-    }
-
-    @Override
     public void setDamage(ItemStack stack, int damage) {
         GearHelper.setDamage(stack, damage, super::setDamage);
     }